requirements.txt
numpy
pandas
matplotlib
opencv-python-headless
ultralytics
keras-cv
tensorflow
requests
Pillow
onnx
